fixed data job loading for updates
